+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 [Full changelog](https://github.com/mozilla/glean.js/compare/v0.24.0...main)
 
+* [#924](https://github.com/mozilla/glean.js/pull/924): Only HTTPS server endpoints outside of testing mode.
+  * In testing mode HTTP may be used. No other protocols are allowed.
+
 # v0.24.0 (2021-11-04)
 
 [Full changelog](https://github.com/mozilla/glean.js/compare/v0.23.0...v0.24.0)